+++ R/zzz.R
@@ -5,6 +5,11 @@ storage_env <- new.env(hash = TRUE)
 		jsonlite::fromJSON(jsonlite::toJSON(data, auto_unbox = TRUE))
 	}, force = TRUE)
   options(SIGMAJS_STORAGE = FALSE)
+  
+  packageStartupMessage(
+    "Welcome to sigmajs\n\n",
+    "Docs: sigmajs.john-coene.com"
+  )
 }
 
 .onLoad <- function(libname, pkgname) {